requirements.txt
openai
requests
beautifulsoup4
docker
selenium
html2text
pyyaml
GitPython